refactor(Header): split into component directory
